Change double to float

+++ Luciol.Plugin/Context/IAnnotation.cs
@@ -17,6 +17,6 @@
         /// <summary>
         /// Value of the annotation
         /// </summary>
-        public double Value { get; }
+        public float Value { get; }
     }
 }
